refactor(cli): extract paper handlers into cli_paper.py (ARCH-2)
Move 30 paper handlers/helpers out of cli.py into new cli_paper.py, plus
relocate _paper_summarize_pending from cli_summarize.py (its only caller,
_paper_command, is a paper handler; cli_paper cannot import from a sibling
cli_* module per rule #1, so the function moves to where its caller lives).
31 real defs total; byte-faithful bodies, no shim.

  handlers (20): _add _remove _move _mark _label _label_bulk _find _verify
    _quarantine _autofill_emit _autofill_apply _fit_check_apply_labels
    _cmd_paper_add_to_cluster _cmd_paper_find _cmd_paper_gaps _paper_command
    _paper_attach_pdfs _paper_enrich_existing _paper_resummarize _paper_upgrade_pdfs
  helpers (10): _find_paper_by_slug_or_doi _display_paper_path
    _read_paper_frontmatter _update_paper_frontmatter _frontmatter_field_text
    _summary_block_has_todo _iter_raw_cluster_dirs _topic_cluster_list
    _manifest_batch_label _parse_bulk_slugs
  + _paper_summarize_pending (from cli_summarize) + _PAPER_FRONTMATTER_RE constant

cli.py 5377 -> 4270 LOC. Selective re-export: cli.py re-imports only the 18
names its dispatch/build_parser reference OR that tests access via
`from research_hub.cli import X` (_cmd_paper_*, _read/_update_paper_frontmatter);
the other 13 are internal to cli_paper (no dead re-exports).

Dependency-sync forwarders (module-level + _sync_cli_dependencies, mirroring
cli_vault): _cli_paper.get_config + ClusterRegistry + require_config. The
require_config forwarder closes a latent SystemExit gap if a future
cli.main(["paper", ...]) test relies on the autouse get_config patch (per
code-reviewer finding).

_read_doi_from_frontmatter left untouched in cli.py: it is uncalled there (a
dead duplicate of cli_clusters' identical copy) and unused by any paper func;
decoupled from this move and tracked for the dedup backlog.

Extraction done Claude-direct (Codex quota-blocked twice; user authorized).
Tests unchanged (pure move). Full suite 3144 passed, 0 failed. Also collapsed
3 runs of 3+ blank lines left by the deletions to PEP8 2-line spacing.

Co-Reviewed-By: code-reviewer subagent (trigger 1: >=3 files / >=50 LOC) -> APPROVE
